requirements.txt
stripe
fastapi
uvicorn
python-dotenv
Jinja2
supabase
python-jose[cryptography]
passlib[bcrypt]